+++ shilyaev.vladislav/A1/circle.hpp
@@ -6,13 +6,13 @@
 
 class Circle: public Shape {
 public:
-  Circle(float radius, point_t pos);
-  float getArea() const;
+  Circle(double radius, point_t pos);
+  double getArea() const;
   rectangle_t getFrameRect() const;
   void move(const point_t& pos);
-  void move(float dx, float dy);
+  void move(double dx, double dy);
 private:
-  float radius_;
+  double radius_;
   point_t pos_;
 };
