+++ shilyaev.vladislav/A1/rectangle.hpp
@@ -6,14 +6,14 @@
 
 class Rectangle: public Shape {
 public:
-  Rectangle(float width, float height, point_t pos);
-  float getArea() const;
+  Rectangle(double width, double height, point_t pos);
+  double getArea() const;
   rectangle_t getFrameRect() const;
   void move(const point_t& pos);
-  void move(float dx, float dy);
+  void move(double dx, double dy);
 private:
-  float width_;
-  float height_;
+  double width_;
+  double height_;
   point_t pos_;
 };
